checktick/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
country-converter==1.0.0
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-reporting==0.21
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flatbuffers==23.3.3
fonttools==4.39.3
freetype-py==2.3.0
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
heroku==0.1.4
idna==3.4
itypes==1.2.0
jax==0.4.6
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.23.5
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
plotly==5.14.1
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==1.5
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
reportlab==4.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
rsa==4.9
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
typing_extensions==4.5.0
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
whitenoise==6.4.0
wrapt==1.14.1
